Extract shared computeHeadingIds helper for anchor validation (#62401)
Copilot-Session: 7bc51b35-b5bd-4cec-9948-a1664d171f4c
